-Added try-catch to all async functions in backend
-Added request data checks (throw error if no required data was passed) in controllers.js

-Finished moving all svgs to SvgIcon, including classNames and props

-Moved to Mongo Atlas to sync up data between computers

-As always, minor bug fixes..

+++ Backend/api/stations/stations.controller.js
@@ -1,30 +1,59 @@
 import { stationsService } from "./stations.service.js"
 
-export async function getAllStations (req, res) {
+export async function getAllStations(req, res) {
+  try {
     const stations = await stationsService.query()
     res.send(stations)
+  } catch (err) {
+    console.log('StationsController: There was an error getting all stations! ', err)
+    throw err
+  }
 }
 
 export async function getStationById(req, res) {
-  const stationId = req.params.stationId
-  var station = await stationsService.getById(stationId)
-  res.send(station)
+  try {
+    const stationId = req.params.stationId
+    if (!stationId) throw 'Cannot get stationId or stationId was not provided (getStationById)'
+    var station = await stationsService.getById(stationId)
+    res.send(station)
+  } catch (err) {
+    console.log('StationsController: There was an error getting requested station! ', err)
+    throw err
+  }
 }
 
-export async function createStation (req, res) {
-  const station = req.body
-  const newStation = await stationsService.addStation(station)
-  res.send(newStation)
+export async function createStation(req, res) {
+  try {
+    const station = req.body
+    if (!station.index || !station.addedBy) throw 'Cannot get index/addedBy or they are not provided (createStation)'
+    const newStation = await stationsService.addStation(station)
+    res.send(newStation)
+  } catch (err) {
+    console.log('StationsController: There was an error creating station! ', err)
+    throw err
+  }
 }
 
-export async function editStation (req,res) {
-  const station = req.body
-  const editedStation = await stationsService.updateStation(station)
-  res.send(editedStation)
+export async function editStation(req, res) {
+  try {
+    const station = req.body
+    if (!station) throw 'Cannot get station or station not provided (editStation)'
+    const editedStation = await stationsService.updateStation(station)
+    res.send(editedStation)
+  } catch (err) {
+    console.log('StationsController: There was an error editing requested station! ', err)
+    throw err
+  }
 }
 
-export async function removeStation (req, res) {
-  const stationId = req.params.stationId
-  await stationsService.deleteStation(stationId)
-  res.send('Deleted Station')
+export async function removeStation(req, res) {
+  try {
+    const stationId = req.params.stationId
+    if (!stationId) throw 'Cannot get stationId or stationId not provided (removeStation)'
+    await stationsService.deleteStation(stationId)
+    res.send('Deleted Station')
+  } catch (err) {
+    console.log('StationsController: There was an error removing requested station! ', err)
+    throw err
+  }
 }
